Fractal: `replace` method
Instead of using `add` et al to add new views to the DOM, `replace` makes sure there's always one instance of a view in existence in a parent.

It adds a new instance if none exist, or updates the state of the last one — and removes all instances other than the last one.